Rename service.models.ModelGenerator
